Unify different versions of dbt manifest parser into one
